# gateway (copy)/docs/swagger.yaml
basePath: /
definitions:
...
  genproto.BudgetCreate:
    properties:
      Amount:
        type: number
      CategoryId:
        type: string
      EndDate:
        type: string
      Period:
        type: string
      StartDate:
        type: string
      UserId:
        type: string
    type: object
  genproto.BudgetGet:
    properties:
      Amount:
        type: number
      CategoryId:
        type: string
      EndDate:
        type: string
      Id:
        type: string
      Period:
        type: string
      StartDate:
        type: string
      UserId:
        type: string
    type: object
  genproto.BudgetList:
    properties:
      Budgets:
        items:
          $ref: '#/definitions/genproto.BudgetGet'
        type: array
      Limit:
        type: integer
      Offset:
        type: integer
      TotalCount:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for Instant Delivery resources
  title: Finance Tracker API Documentation
  version: "1.0"
paths:
...
  /budgets/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a budget by its ID
      parameters:
      - description: Budget ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Budget details
          schema:
            $ref: '#/definitions/genproto.BudgetGet'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Budget not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Budget
      tags:
      - Budget
    put:
      consumes:
      - application/json
      description: Update an existing budget
      parameters:
      - description: Budget ID
        in: path
        name: id
        required: true
        type: string
      - description: Budget data
        in: body
        name: budget
        required: true
        schema:
          $ref: '#/definitions/genproto.BudgetCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Budget updated successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Budget
      tags:
      - Budget
  /budgets/create:
    post:
      consumes:
      - application/json
      description: Create a new budget
      parameters:
      - description: Budget data
        in: body
        name: budget
        required: true
        schema:
          $ref: '#/definitions/genproto.BudgetCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Budget created successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Budget
      tags:
      - Budget
  /budgets/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a budget by its ID
      parameters:
      - description: Budget ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Budget deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Budget not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Budget
      tags:
      - Budget
  /budgets/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of budgets
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Category ID
        in: query
        name: category_id
        type: string
      - description: Minimum Amount
        in: query
        name: amount_from
        type: number
      - description: Maximum Amount
        in: query
        name: amount_to
        type: number
      - description: Budget Period
        in: query
        name: period
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of budgets
          schema:
            $ref: '#/definitions/genproto.BudgetList'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List Budgets
      tags:
      - Budget
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"